refactor: main page 파일 위치 이동
